requirements.txt
aiogram>=3.4,<4.0
fastapi>=0.111,<1.0
uvicorn>=0.30,<1.0
python-dotenv>=1.0,<2.0
python-multipart>=0.0.9,<1.0
opencv-python-headless>=4.10,<5.0
httpx>=0.27,<1.0
